+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # Buxfixes
 
-- Made it possible to retrieve banners from shields crafted without the mod (WARNING: You will just get a banner with the patterns, no custom names or any other components)
-- Shields will now stay in the slot they are when using the shield on a cauldron.
+- Made the mod server-only at the cost of not retaining the original banner components.